fix: remove automaton configs, fix readme, refactor, upgrade hardhat version
