--- config_samples/optimism/automaton/automaton_config_L2.json
@@ -1,85 +1,0 @@
-{
-  "contracts": {
-    "0xB34F2747BCd9BCC4107A0ccEb43D5dcdd7Fabf89": "OssifiableProxy",
-    "0xa989A4B3A26e28DC9d106F163B2B1f35153E0517": "TokenRateOracle",
-    "0x24B47cd3A74f1799b32B2de11073764Cb1bb318B": "OssifiableProxy",
-    "0x298953B9426eba4F35a137a4754278a16d97A063": "ERC20BridgedPermit",
-    "0xf49D208B5C7b10415C7BeAFe9e656F2DF9eDfe3B": "OssifiableProxy",
-    "0xFd21C82c99ddFa56EB0B9B2D1d0709b7E26D1B2C": "ERC20RebasableBridgedPermit",
-    "0xdBA2760246f315203F8B716b3a7590F0FFdc704a": "OssifiableProxy",
-    "0xD48c69358193a34aC035ea7dfB70daDea1600112": "L2ERC20ExtendedTokensBridge"
-  },
-  "explorer_token_env_var": "OPTISCAN_EXPLORER_TOKEN",
-  "explorer_hostname_env_var": "L2_EXPLORER_API_HOSTNAME",
-  "github_repo": {
-    "url": "https://github.com/lidofinance/lido-l2-with-steth",
-    "commit": "8f19e1101a211c8f3d42af7ffcb87ab0ebcf750c",
-    "relative_root": ""
-  },
-  "dependencies": {
-    "@openzeppelin/contracts": {
-      "url": "https://github.com/OpenZeppelin/openzeppelin-contracts",
-      "commit": "d4fb3a89f9d0a39c7ee6f2601d33ffbf30085322",
-      "relative_root": "contracts"
-    }
-  },
-  "fail_on_comparison_error": true,
-  "bytecode_comparison": {
-    "constructor_args": {
-      "0xB34F2747BCd9BCC4107A0ccEb43D5dcdd7Fabf89": [
-        "0xB34F2747BCd9BCC4107A0ccEb43D5dcdd7Fabf89",
-        "0xB34F2747BCd9BCC4107A0ccEb43D5dcdd7Fabf89",
-        "0x"
-      ],
-      "0xa989A4B3A26e28DC9d106F163B2B1f35153E0517": [
-        "0x4200000000000000000000000000000000000007",
-        "0xdBA2760246f315203F8B716b3a7590F0FFdc704a",
-        "0x4067B05a6B2f6801Bfb8d4fF417eD32e71c216d9",
-        86400,
-        172800,
-        500,
-        86400,
-        3600
-      ],
-      "0x24B47cd3A74f1799b32B2de11073764Cb1bb318B": [
-        "0x24B47cd3A74f1799b32B2de11073764Cb1bb318B",
-        "0x24B47cd3A74f1799b32B2de11073764Cb1bb318B",
-        "0x"
-      ],
-      "0x298953B9426eba4F35a137a4754278a16d97A063": [
-        "Wrapped liquid staked Ether 2.0",
-        "wstETH",
-        "2",
-        18,
-        "0xdBA2760246f315203F8B716b3a7590F0FFdc704a"
-      ],
-      "0xf49D208B5C7b10415C7BeAFe9e656F2DF9eDfe3B": [
-        "0xf49D208B5C7b10415C7BeAFe9e656F2DF9eDfe3B",
-        "0xf49D208B5C7b10415C7BeAFe9e656F2DF9eDfe3B",
-        "0x"
-      ],
-      "0xFd21C82c99ddFa56EB0B9B2D1d0709b7E26D1B2C": [
-        "Liquid staked Ether 2.0",
-        "stETH",
-        "1",
-        18,
-        "0x24B47cd3A74f1799b32B2de11073764Cb1bb318B",
-        "0xB34F2747BCd9BCC4107A0ccEb43D5dcdd7Fabf89",
-        "0xdBA2760246f315203F8B716b3a7590F0FFdc704a"
-      ],
-      "0xdBA2760246f315203F8B716b3a7590F0FFdc704a": [
-        "0xdBA2760246f315203F8B716b3a7590F0FFdc704a",
-        "0xdBA2760246f315203F8B716b3a7590F0FFdc704a",
-        "0x"
-      ],
-      "0xD48c69358193a34aC035ea7dfB70daDea1600112": [
-        "0x4200000000000000000000000000000000000007",
-        "0x4Abf633d9c0F4aEebB4C2E3213c7aa1b8505D332",
-        "0xB82381A3fBD3FaFA77B3a7bE693342618240067b",
-        "0x3e3FE7dBc6B4C189E7128855dD526361c49b40Af",
-        "0x24B47cd3A74f1799b32B2de11073764Cb1bb318B",
-        "0xf49D208B5C7b10415C7BeAFe9e656F2DF9eDfe3B"
-      ]
-    }
-  }
-}
