--- config_samples/optimism/automaton/automaton_config_L2_gov.json
@@ -1,34 +1,0 @@
-{
-  "contracts": {
-    "0xf695357C66bA514150Da95b189acb37b46DDe602": "OptimismBridgeExecutor"
-  },
-  "explorer_token_env_var": "OPTISCAN_EXPLORER_TOKEN",
-  "explorer_hostname_env_var": "L2_EXPLORER_API_HOSTNAME",
-  "github_repo": {
-    "url": "https://github.com/lidofinance/governance-crosschain-bridges",
-    "commit": "8fa25b0080dd3dcc2390313631aea6796a12c9d8",
-    "relative_root": ""
-  },
-  "dependencies": {
-    "@openzeppelin/contracts": {
-      "url": "https://github.com/OpenZeppelin/openzeppelin-contracts",
-      "commit": "d4fb3a89f9d0a39c7ee6f2601d33ffbf30085322",
-      "relative_root": "contracts",
-      "//": "OZ 4.6.0"
-    }
-  },
-  "fail_on_comparison_error": true,
-  "bytecode_comparison": {
-    "constructor_args": {
-      "0xf695357C66bA514150Da95b189acb37b46DDe602": [
-        "0x4200000000000000000000000000000000000007",
-        "0x32A0E5828B62AAb932362a4816ae03b860b65e83",
-        0,
-        86400,
-        0,
-        1,
-        "0x0000000000000000000000000000000000000000"
-      ]
-    }
-  }
-}
