+++ package.json
@@ -8,7 +8,7 @@
     "@commitlint/config-conventional": "^19.5.0"
   },
   "dependencies": {
-    "hardhat": "^2.22.15",
+    "hardhat": "^2.22.17",
     "kill-port": "^2.0.1"
   },
   "packageManager": "npm@10.8.2",
